roomGroup init & seperate utilities
